Crate raytracer_5607

Modules

camera
A virtual camera

image
Represents a collection of pixels (an image)

intersection
Intersection record

lights
Various lights that can be represented in a ray tracer

material
A material to apply to a ray-traced object

pixel
Represents a pixel

ray
A half-line (a starting point and a direction)

ray_tracer
The main ray tracing implementation

scene
A simple description of a scene used for ray tracing

sphere
Representation of a sphere to be ray traced

vector
A simple vector