Struct raytracer_5607::pixel::Pixel

pub struct Pixel {
    pub r: f64,
    pub g: f64,
    pub b: f64,
    pub a: f64,
    pub with_clamping: bool,
}

Holds all pixel information as a float

Fields
r: f64 g: f64 b: f64 a: f64 with_clamping: bool
Methods

impl Pixel

pub fn from_slice_unclamped(numbers: &[f64]) -> Self

pub fn from_pix_unclamped(pix: Self) -> Self

pub fn from_rgba_unclamped(r: f64, g: f64, b: f64, a: f64) -> Self

Construct a new pixel from R, G, B, A values, but don't clamp values

pub fn from_rgba(r: f64, g: f64, b: f64, a: f64) -> Self

Construct a new pixel from R, G, B, A values

pub fn from_rgb(r: f64, g: f64, b: f64) -> Self

Construct a new pixel from R, G, B values

pub fn clamp(self) -> Self

Construct a new pixel with all values clamped, consuming self

pub fn set_rgba(&mut self, r: f64, g: f64, b: f64, a: f64)

Set RGBA components

pub fn set_rgb(&mut self, r: f64, g: f64, b: f64)

Set RGB components

pub fn luminance(&self) -> f64

Luminance of a pixel Luminance formula from Wikipedia

pub fn lerp(&self, other: &Pixel, amount: f64) -> Pixel

Linearly interpolate with another pixel

pub fn quant(&self, num_bins: usize) -> Pixel

Quantize the bits of this pixel into bins